camera_upload_dev merge develop-studio

+++ app/src/main/java/mega/privacy/android/app/UploadService.java
@@ -73,8 +73,6 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 	private boolean isForeground = false;
 	private boolean canceled;
 
-	private boolean isQRFile = false;
-
 	MegaApplication app;
 	MegaApiAndroid megaApi;
 	MegaChatApiAndroid megaChatApi;
@@ -197,7 +195,7 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 		log("onHandleIntent");
 
 		final File file = new File(intent.getStringExtra(EXTRA_FILEPATH));
-		isQRFile = intent.getBooleanExtra("qrfile", false);
+
 		if(file!=null){
 			log("File to manage: "+file.getAbsolutePath());
 		}
@@ -236,7 +234,6 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 					}
 					case CHECK_FILE_TO_UPLOAD_COPY: {
 						log("CHECK_FILE_TO_UPLOAD_COPY");
-						copiedCount++;
 						break;
 					}
 					case CHECK_FILE_TO_UPLOAD_SAME_FILE_IN_FOLDER: {
@@ -278,7 +275,6 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 					}
 					case CHECK_FILE_TO_UPLOAD_COPY: {
 						log("CHECK_FILE_TO_UPLOAD_COPY");
-						copiedCount++;
 						break;
 					}
 					case CHECK_FILE_TO_UPLOAD_OVERWRITE: {
@@ -859,15 +855,19 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 					}
 				}
 
-				if (getApplicationContext().getExternalCacheDir() != null && !isQRFile) {
-					File localFile = new File(getApplicationContext().getExternalCacheDir(), transfer.getFileName());
-					if (localFile.exists()) {
+				String qrFileName = megaApi.getMyEmail() + "QRcode.jpg";
+
+				if (getApplicationContext().getExternalCacheDir() != null) {
+					File qrDir = new File (getApplicationContext().getExternalCacheDir(), "qrMEGA");
+					File localFile = new File(qrDir, transfer.getFileName());
+					if (localFile.exists() && !localFile.getName().equals(qrFileName)) {
 						log("Delete file!: " + localFile.getAbsolutePath());
 						localFile.delete();
 					}
-				} else if (!isQRFile){
-					File localFile = new File(getApplicationContext().getCacheDir(), transfer.getFileName());
-					if (localFile.exists()) {
+				} else {
+					File qrDir = getApplicationContext().getDir("qrMEGA", 0);
+					File localFile = new File(qrDir, transfer.getFileName());
+					if (localFile.exists() && !localFile.getName().equals(qrFileName)) {
 						log("Delete file!: " + localFile.getAbsolutePath());
 						localFile.delete();
 					}
@@ -888,8 +888,7 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 					log("transfer.getPath() is NULL");
 				}
 
-				int total = totalUploadsCompleted + copiedCount;
-				if (total==totalUploads && transfersCount == 0) {
+				if (transfersCopy.isEmpty() && totalUploadsCompleted==totalUploads && transfersCount == 0) {
 					onQueueComplete();
 				}
 				else{
@@ -1000,6 +999,7 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 		if (request.getType() == MegaRequest.TYPE_COPY){
 			log("TYPE_COPY finished");
 			if (e.getErrorCode() == MegaError.API_OK){
+				copiedCount++;
 				MegaNode n = megaApi.getNodeByHandle(request.getNodeHandle());
 				if (n != null){
 					String currentNodeName = n.getName();
@@ -1012,9 +1012,12 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 						}
 					}
 					transfersCopy.remove(megaFingerPrint);
-					int total = totalUploadsCompleted + copiedCount;
-					if (total==totalUploads){
-						onQueueComplete();
+
+					if(transfersCopy.isEmpty()){
+						//int total = totalUploadsCompleted + copiedCount;
+						if (totalUploads==totalUploadsCompleted && transfersCount == 0) {
+							onQueueComplete();
+						}
 					}
 				}
 				else{
